CHAPTER_I/Exams_REST_API/requirements.txt
Django==2.1.5
django-cors-headers==2.1.0
djangorestframework==3.9.1
drf-extensions==0.4.0
pkg-resources==0.0.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
